refactor(web): unify workspace routes — drop tab-driven URL subpaths
Per the follow-up on #467, the workspace route tree is now just two
files: `index.tsx` (project list at `/`) and `workspace.$workspaceId.tsx`
(workspace view at `/workspace/$id`). The five tab-specific child routes
(`/changes`, `/code`, `/code/$`, `/code/`, `/terminal`) are gone, and the
mobile workspace tab nav is now driven by local React state instead of
URL navigation.

Why
---

Post the shared-dockview refactor (#456 / PR #461) the desktop layout
ignored these child routes entirely — every panel renders once at the
AppShell-level dockview. On mobile they encoded the active tab in the
URL, which meant every tab tap pushed onto the browser history and the
`band-tab:` sessionStorage entry tried to "restore" the last tab on
return visits. Both behaviours added complexity without clear UX wins.

What changed
------------

* `workspace.$workspaceId.tsx` now owns:
  - Local `activeTab` (defaults to "chat" on every workspace open — no
    persistence)
  - Local `currentFile` for the Files tab
  - Inline rendering of `ChatView` / `DiffView` / `CodeBrowserView` /
    `DockviewTerminalContainer` per active tab (conditional render,
    matching the pre-refactor unmount-on-switch behaviour)
* `WorkspaceTabNav` is button-driven only; the `tabHrefs` prop is gone
* `WebCapabilities.getWorkspaceHref` always returns `/workspace/$id` —
  the `VALID_TAB_PREFIXES` + sessionStorage `band-tab:` mechanism is
  deleted
* `dispatch-open-file.ts` keeps the same handler interface; the mobile
  implementations in `__root.tsx` now queue a tab-switch + optional file
  via `setMobilePendingAction` before navigating to `/workspace/$id`.
  `MobileWorkspaceLayout` drains this queue in its mount effect (same
  Map + listener-set pattern as `pending-external-open.ts`)
* `routeTree.gen.ts` shrinks from ~190 lines covering 7 routes to ~85
  lines covering 2
* Three e2e tests that deep-linked to `/workspace/$id/code` were updated
  to use `/workspace/$id`; `find-in-markdown-preview` now drives the UI
  flow (click Files tab → click `GUIDE.md` in the tree) instead of URL
  deep-linking

Net diff: -731 / +317 lines. 852 server tests + 32 e2e + 194 desktop
tests all pass; lint clean; knip clean.

Closes #467.
