+++ apps/web/src/routes/__root.tsx
@@ -35,6 +35,7 @@ import { useZoom } from "../hooks/useZoom";
 import { getElectronBridge } from "../lib/desktop-ipc";
 import { dispatchOpenFileEvent } from "../lib/dispatch-open-file";
 import { isDesktop } from "../lib/is-desktop";
+import { setMobilePendingAction } from "../lib/mobile-pending-action";
 import { parseWorkspaceFromPath } from "../lib/parse-workspace";
 import {
   applyZoomLevel,
@@ -351,19 +352,30 @@ function AppShell() {
         handlers: {
           onOpenFile: crossPanelHandlers.onOpenFile,
           onActivateFilesPanel: crossPanelHandlers.onActivateFilesPanel,
-          navigateInWorkspace: (workspaceId, filePath) =>
+          // Mobile: the workspace URL no longer carries the active tab or
+          // selected file — both live in `MobileWorkspaceLayout`'s local
+          // state. We queue the tab-switch + file open through
+          // `setMobilePendingAction` BEFORE navigating, so the soon-to-be-
+          // mounted layout drains the queue in its `useEffect` (the queue
+          // module mirrors `pending-external-open.ts` — same pattern as
+          // CodeBrowserView's external-file drain). See issue #467.
+          navigateInWorkspace: (workspaceId, filePath) => {
+            setMobilePendingAction(workspaceId, { tab: "code", filePath });
             router.navigate({
-              to: "/workspace/$workspaceId/code/$",
-              params: {
-                workspaceId: encodeURIComponent(workspaceId),
-                _splat: filePath,
-              },
-            }),
-          navigateToWorkspaceCode: (workspaceId) =>
-            router.navigate({
-              to: "/workspace/$workspaceId/code",
+              to: "/workspace/$workspaceId",
               params: { workspaceId: encodeURIComponent(workspaceId) },
-            }),
+            });
+          },
+          navigateToWorkspaceCode: (workspaceId) => {
+            // External path: the actual file lives in
+            // `pending-external-open` and is drained by `CodeBrowserView`'s
+            // own mount effect — we just need to switch to the Files tab.
+            setMobilePendingAction(workspaceId, { tab: "code" });
+            router.navigate({
+              to: "/workspace/$workspaceId",
+              params: { workspaceId: encodeURIComponent(workspaceId) },
+            });
+          },
         },
       });
     });
